--- pkg/audit/common.go
@@ -1,12 +1,0 @@
-package audit
-
-import (
-	"time"
-)
-
-type AuthenticatedClientMeta struct {
-	ClientIds     []string
-	TokenIssuedAt time.Time
-	ClientIP      string
-	Subject       string
-}
