+++ pkg/rpc/adminservice/attributes.go
@@ -2,9 +2,6 @@ package adminservice
 
 import (
 	"context"
-	"time"
-
-	"github.com/flyteorg/flyteadmin/pkg/audit"
 
 	"github.com/flyteorg/flyteadmin/pkg/rpc/adminservice/util"
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
@@ -15,7 +12,6 @@ import (
 func (m *AdminService) UpdateWorkflowAttributes(ctx context.Context, request *admin.WorkflowAttributesUpdateRequest) (
 	*admin.WorkflowAttributesUpdateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -24,16 +20,6 @@ func (m *AdminService) UpdateWorkflowAttributes(ctx context.Context, request *ad
 	m.Metrics.workflowAttributesEndpointMetrics.update.Time(func() {
 		response, err = m.ResourceManager.UpdateWorkflowAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateWorkflowAttributes",
-		map[string]string{
-			audit.Project: request.Attributes.Project,
-			audit.Domain:  request.Attributes.Domain,
-			audit.Name:    request.Attributes.Workflow,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.workflowAttributesEndpointMetrics.update)
 	}
@@ -44,7 +30,6 @@ func (m *AdminService) UpdateWorkflowAttributes(ctx context.Context, request *ad
 func (m *AdminService) GetWorkflowAttributes(ctx context.Context, request *admin.WorkflowAttributesGetRequest) (
 	*admin.WorkflowAttributesGetResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -53,16 +38,6 @@ func (m *AdminService) GetWorkflowAttributes(ctx context.Context, request *admin
 	m.Metrics.workflowAttributesEndpointMetrics.get.Time(func() {
 		response, err = m.ResourceManager.GetWorkflowAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetWorkflowAttributes",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-			audit.Name:    request.Workflow,
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.workflowAttributesEndpointMetrics.get)
 	}
@@ -73,7 +48,6 @@ func (m *AdminService) GetWorkflowAttributes(ctx context.Context, request *admin
 func (m *AdminService) DeleteWorkflowAttributes(ctx context.Context, request *admin.WorkflowAttributesDeleteRequest) (
 	*admin.WorkflowAttributesDeleteResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -82,16 +56,6 @@ func (m *AdminService) DeleteWorkflowAttributes(ctx context.Context, request *ad
 	m.Metrics.workflowAttributesEndpointMetrics.delete.Time(func() {
 		response, err = m.ResourceManager.DeleteWorkflowAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"DeleteWorkflowAttributes",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-			audit.Name:    request.Workflow,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.workflowAttributesEndpointMetrics.delete)
 	}
@@ -102,7 +66,6 @@ func (m *AdminService) DeleteWorkflowAttributes(ctx context.Context, request *ad
 func (m *AdminService) UpdateProjectDomainAttributes(ctx context.Context, request *admin.ProjectDomainAttributesUpdateRequest) (
 	*admin.ProjectDomainAttributesUpdateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -111,15 +74,6 @@ func (m *AdminService) UpdateProjectDomainAttributes(ctx context.Context, reques
 	m.Metrics.projectDomainAttributesEndpointMetrics.update.Time(func() {
 		response, err = m.ResourceManager.UpdateProjectDomainAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateProjectDomainAttributes",
-		map[string]string{
-			audit.Project: request.Attributes.Project,
-			audit.Domain:  request.Attributes.Domain,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.projectDomainAttributesEndpointMetrics.update)
 	}
@@ -130,7 +84,6 @@ func (m *AdminService) UpdateProjectDomainAttributes(ctx context.Context, reques
 func (m *AdminService) GetProjectDomainAttributes(ctx context.Context, request *admin.ProjectDomainAttributesGetRequest) (
 	*admin.ProjectDomainAttributesGetResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -139,15 +92,6 @@ func (m *AdminService) GetProjectDomainAttributes(ctx context.Context, request *
 	m.Metrics.workflowAttributesEndpointMetrics.get.Time(func() {
 		response, err = m.ResourceManager.GetProjectDomainAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetProjectDomainAttributes",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.workflowAttributesEndpointMetrics.get)
 	}
@@ -158,7 +102,6 @@ func (m *AdminService) GetProjectDomainAttributes(ctx context.Context, request *
 func (m *AdminService) DeleteProjectDomainAttributes(ctx context.Context, request *admin.ProjectDomainAttributesDeleteRequest) (
 	*admin.ProjectDomainAttributesDeleteResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -167,15 +110,6 @@ func (m *AdminService) DeleteProjectDomainAttributes(ctx context.Context, reques
 	m.Metrics.workflowAttributesEndpointMetrics.delete.Time(func() {
 		response, err = m.ResourceManager.DeleteProjectDomainAttributes(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"DeleteProjectDomainAttributes",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.workflowAttributesEndpointMetrics.delete)
 	}
@@ -186,7 +120,6 @@ func (m *AdminService) DeleteProjectDomainAttributes(ctx context.Context, reques
 func (m *AdminService) ListMatchableAttributes(ctx context.Context, request *admin.ListMatchableAttributesRequest) (
 	*admin.ListMatchableAttributesResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -195,14 +128,6 @@ func (m *AdminService) ListMatchableAttributes(ctx context.Context, request *adm
 	m.Metrics.matchableAttributesEndpointMetrics.list.Time(func() {
 		response, err = m.ResourceManager.ListAll(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListMatchableAttributes",
-		map[string]string{
-			audit.ResourceType: request.ResourceType.String(),
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.matchableAttributesEndpointMetrics.list)
 	}
