+++ pkg/rpc/adminservice/execution.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"time"
 
-	"github.com/flyteorg/flyteadmin/pkg/audit"
-
 	"github.com/flyteorg/flyteadmin/pkg/rpc/adminservice/util"
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
 	"google.golang.org/grpc/codes"
@@ -24,16 +22,6 @@ func (m *AdminService) CreateExecution(
 	m.Metrics.executionEndpointMetrics.create.Time(func() {
 		response, err = m.ExecutionManager.CreateExecution(ctx, *request, requestedAt)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ExecutionCreateRequest",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-			audit.Name:    request.Name,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.create)
 	}
@@ -53,12 +41,6 @@ func (m *AdminService) RelaunchExecution(
 	m.Metrics.executionEndpointMetrics.relaunch.Time(func() {
 		response, err = m.ExecutionManager.RelaunchExecution(ctx, *request, requestedAt)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ExecutionCreateRequest",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.relaunch)
 	}
@@ -78,12 +60,6 @@ func (m *AdminService) RecoverExecution(
 	m.Metrics.executionEndpointMetrics.recover.Time(func() {
 		response, err = m.ExecutionManager.RecoverExecution(ctx, *request, requestedAt)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ExecutionCreateRequest",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.relaunch)
 	}
@@ -94,7 +70,6 @@ func (m *AdminService) RecoverExecution(
 func (m *AdminService) CreateWorkflowEvent(
 	ctx context.Context, request *admin.WorkflowExecutionEventRequest) (*admin.WorkflowExecutionEventResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -103,12 +78,6 @@ func (m *AdminService) CreateWorkflowEvent(
 	m.Metrics.executionEndpointMetrics.createEvent.Time(func() {
 		response, err = m.ExecutionManager.CreateWorkflowEvent(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"CreateWorkflowEvent",
-		audit.ParametersFromExecutionIdentifier(request.Event.ExecutionId),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.createEvent)
 	}
@@ -120,7 +89,6 @@ func (m *AdminService) CreateWorkflowEvent(
 func (m *AdminService) GetExecution(
 	ctx context.Context, request *admin.WorkflowExecutionGetRequest) (*admin.Execution, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -129,12 +97,6 @@ func (m *AdminService) GetExecution(
 	m.Metrics.executionEndpointMetrics.get.Time(func() {
 		response, err = m.ExecutionManager.GetExecution(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetExecution",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.get)
 	}
@@ -154,12 +116,6 @@ func (m *AdminService) UpdateExecution(
 	m.Metrics.executionEndpointMetrics.update.Time(func() {
 		response, err = m.ExecutionManager.UpdateExecution(ctx, *request, requestedAt)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateExecution",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.update)
 	}
@@ -170,7 +126,6 @@ func (m *AdminService) UpdateExecution(
 func (m *AdminService) GetExecutionData(
 	ctx context.Context, request *admin.WorkflowExecutionGetDataRequest) (*admin.WorkflowExecutionGetDataResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -179,12 +134,6 @@ func (m *AdminService) GetExecutionData(
 	m.Metrics.executionEndpointMetrics.get.Time(func() {
 		response, err = m.ExecutionManager.GetExecutionData(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetExecutionData",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.getData)
 	}
@@ -195,7 +144,6 @@ func (m *AdminService) GetExecutionData(
 func (m *AdminService) ListExecutions(
 	ctx context.Context, request *admin.ResourceListRequest) (*admin.ExecutionList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -204,12 +152,6 @@ func (m *AdminService) ListExecutions(
 	m.Metrics.executionEndpointMetrics.list.Time(func() {
 		response, err = m.ExecutionManager.ListExecutions(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListExecutions",
-		audit.ParametersFromNamedEntityIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.list)
 	}
@@ -220,7 +162,6 @@ func (m *AdminService) ListExecutions(
 func (m *AdminService) TerminateExecution(
 	ctx context.Context, request *admin.ExecutionTerminateRequest) (*admin.ExecutionTerminateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -229,12 +170,6 @@ func (m *AdminService) TerminateExecution(
 	m.Metrics.executionEndpointMetrics.terminate.Time(func() {
 		response, err = m.ExecutionManager.TerminateExecution(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"TerminateExecution",
-		audit.ParametersFromExecutionIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.executionEndpointMetrics.terminate)
 	}
