+++ pkg/rpc/adminservice/launch_plan.go
@@ -2,9 +2,6 @@ package adminservice
 
 import (
 	"context"
-	"time"
-
-	"github.com/flyteorg/flyteadmin/pkg/audit"
 
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/core"
 	"github.com/flyteorg/flytestdlib/logger"
@@ -18,7 +15,6 @@ import (
 func (m *AdminService) CreateLaunchPlan(
 	ctx context.Context, request *admin.LaunchPlanCreateRequest) (*admin.LaunchPlanCreateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -27,12 +23,6 @@ func (m *AdminService) CreateLaunchPlan(
 	m.Metrics.launchPlanEndpointMetrics.create.Time(func() {
 		response, err = m.LaunchPlanManager.CreateLaunchPlan(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"CreateLaunchPlan",
-		audit.ParametersFromIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.create)
 	}
@@ -42,7 +32,6 @@ func (m *AdminService) CreateLaunchPlan(
 
 func (m *AdminService) GetLaunchPlan(ctx context.Context, request *admin.ObjectGetRequest) (*admin.LaunchPlan, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -57,12 +46,6 @@ func (m *AdminService) GetLaunchPlan(ctx context.Context, request *admin.ObjectG
 	m.Metrics.launchPlanEndpointMetrics.get.Time(func() {
 		response, err = m.LaunchPlanManager.GetLaunchPlan(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetLaunchPlan",
-		audit.ParametersFromIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.get)
 	}
@@ -73,7 +56,6 @@ func (m *AdminService) GetLaunchPlan(ctx context.Context, request *admin.ObjectG
 
 func (m *AdminService) GetActiveLaunchPlan(ctx context.Context, request *admin.ActiveLaunchPlanRequest) (*admin.LaunchPlan, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -82,12 +64,6 @@ func (m *AdminService) GetActiveLaunchPlan(ctx context.Context, request *admin.A
 	m.Metrics.launchPlanEndpointMetrics.getActive.Time(func() {
 		response, err = m.LaunchPlanManager.GetActiveLaunchPlan(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetActiveLaunchPlan",
-		audit.ParametersFromNamedEntityIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.getActive)
 	}
@@ -98,7 +74,6 @@ func (m *AdminService) GetActiveLaunchPlan(ctx context.Context, request *admin.A
 func (m *AdminService) UpdateLaunchPlan(ctx context.Context, request *admin.LaunchPlanUpdateRequest) (
 	*admin.LaunchPlanUpdateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -113,12 +88,6 @@ func (m *AdminService) UpdateLaunchPlan(ctx context.Context, request *admin.Laun
 	m.Metrics.launchPlanEndpointMetrics.update.Time(func() {
 		response, err = m.LaunchPlanManager.UpdateLaunchPlan(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateLaunchPlan",
-		audit.ParametersFromIdentifier(request.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.update)
 	}
@@ -129,7 +98,6 @@ func (m *AdminService) UpdateLaunchPlan(ctx context.Context, request *admin.Laun
 func (m *AdminService) ListLaunchPlans(ctx context.Context, request *admin.ResourceListRequest) (
 	*admin.LaunchPlanList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Empty request.  Please rephrase.")
 	}
@@ -138,12 +106,6 @@ func (m *AdminService) ListLaunchPlans(ctx context.Context, request *admin.Resou
 	m.Metrics.launchPlanEndpointMetrics.list.Time(func() {
 		response, err = m.LaunchPlanManager.ListLaunchPlans(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListLaunchPlans",
-		audit.ParametersFromNamedEntityIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.list)
 	}
@@ -155,7 +117,6 @@ func (m *AdminService) ListLaunchPlans(ctx context.Context, request *admin.Resou
 func (m *AdminService) ListActiveLaunchPlans(ctx context.Context, request *admin.ActiveLaunchPlanListRequest) (
 	*admin.LaunchPlanList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Empty request.  Please rephrase.")
 	}
@@ -164,15 +125,6 @@ func (m *AdminService) ListActiveLaunchPlans(ctx context.Context, request *admin
 	m.Metrics.launchPlanEndpointMetrics.listActive.Time(func() {
 		response, err = m.LaunchPlanManager.ListActiveLaunchPlans(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListActiveLaunchPlans",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.listActive)
 	}
@@ -184,7 +136,6 @@ func (m *AdminService) ListActiveLaunchPlans(ctx context.Context, request *admin
 func (m *AdminService) ListLaunchPlanIds(ctx context.Context, request *admin.NamedEntityIdentifierListRequest) (
 	*admin.NamedEntityIdentifierList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Empty request.  Please rephrase.")
 	}
@@ -194,15 +145,6 @@ func (m *AdminService) ListLaunchPlanIds(ctx context.Context, request *admin.Nam
 	m.Metrics.launchPlanEndpointMetrics.listIds.Time(func() {
 		response, err = m.LaunchPlanManager.ListLaunchPlanIds(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListLaunchPlanIds",
-		map[string]string{
-			audit.Project: request.Project,
-			audit.Domain:  request.Domain,
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.launchPlanEndpointMetrics.listIds)
 	}
