+++ pkg/rpc/adminservice/named_entity.go
@@ -2,9 +2,6 @@ package adminservice
 
 import (
 	"context"
-	"time"
-
-	"github.com/flyteorg/flyteadmin/pkg/audit"
 
 	"github.com/flyteorg/flyteadmin/pkg/rpc/adminservice/util"
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
@@ -14,7 +11,6 @@ import (
 
 func (m *AdminService) GetNamedEntity(ctx context.Context, request *admin.NamedEntityGetRequest) (*admin.NamedEntity, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -24,12 +20,6 @@ func (m *AdminService) GetNamedEntity(ctx context.Context, request *admin.NamedE
 	m.Metrics.namedEntityEndpointMetrics.get.Time(func() {
 		response, err = m.NamedEntityManager.GetNamedEntity(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetNamedEntity",
-		audit.ParametersFromNamedEntityIdentifierAndResource(request.Id, request.ResourceType),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.namedEntityEndpointMetrics.get)
 	}
@@ -41,7 +31,6 @@ func (m *AdminService) GetNamedEntity(ctx context.Context, request *admin.NamedE
 func (m *AdminService) UpdateNamedEntity(ctx context.Context, request *admin.NamedEntityUpdateRequest) (
 	*admin.NamedEntityUpdateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -51,12 +40,6 @@ func (m *AdminService) UpdateNamedEntity(ctx context.Context, request *admin.Nam
 	m.Metrics.namedEntityEndpointMetrics.update.Time(func() {
 		response, err = m.NamedEntityManager.UpdateNamedEntity(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateNamedEntity",
-		audit.ParametersFromNamedEntityIdentifierAndResource(request.Id, request.ResourceType),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.namedEntityEndpointMetrics.update)
 	}
@@ -67,7 +50,6 @@ func (m *AdminService) UpdateNamedEntity(ctx context.Context, request *admin.Nam
 func (m *AdminService) ListNamedEntities(ctx context.Context, request *admin.NamedEntityListRequest) (
 	*admin.NamedEntityList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -77,16 +59,6 @@ func (m *AdminService) ListNamedEntities(ctx context.Context, request *admin.Nam
 	m.Metrics.namedEntityEndpointMetrics.list.Time(func() {
 		response, err = m.NamedEntityManager.ListNamedEntities(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListNamedEntities",
-		map[string]string{
-			audit.Project:      request.Project,
-			audit.Domain:       request.Domain,
-			audit.ResourceType: request.ResourceType.String(),
-		},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.namedEntityEndpointMetrics.list)
 	}
