+++ pkg/rpc/adminservice/node_execution.go
@@ -2,9 +2,6 @@ package adminservice
 
 import (
 	"context"
-	"time"
-
-	"github.com/flyteorg/flyteadmin/pkg/audit"
 
 	"github.com/flyteorg/flytestdlib/logger"
 
@@ -18,7 +15,6 @@ import (
 func (m *AdminService) CreateNodeEvent(
 	ctx context.Context, request *admin.NodeExecutionEventRequest) (*admin.NodeExecutionEventResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -27,12 +23,6 @@ func (m *AdminService) CreateNodeEvent(
 	m.Metrics.nodeExecutionEndpointMetrics.createEvent.Time(func() {
 		response, err = m.NodeExecutionManager.CreateNodeEvent(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"CreateNodeEvent",
-		audit.ParametersFromNodeExecutionIdentifier(request.Event.Id),
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.nodeExecutionEndpointMetrics.createEvent)
 	}
@@ -43,7 +33,6 @@ func (m *AdminService) CreateNodeEvent(
 func (m *AdminService) GetNodeExecution(
 	ctx context.Context, request *admin.NodeExecutionGetRequest) (*admin.NodeExecution, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -52,12 +41,6 @@ func (m *AdminService) GetNodeExecution(
 	m.Metrics.nodeExecutionEndpointMetrics.get.Time(func() {
 		response, err = m.NodeExecutionManager.GetNodeExecution(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetNodeExecution",
-		audit.ParametersFromNodeExecutionIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.nodeExecutionEndpointMetrics.get)
 	}
@@ -68,7 +51,6 @@ func (m *AdminService) GetNodeExecution(
 func (m *AdminService) ListNodeExecutions(
 	ctx context.Context, request *admin.NodeExecutionListRequest) (*admin.NodeExecutionList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -77,12 +59,6 @@ func (m *AdminService) ListNodeExecutions(
 	m.Metrics.nodeExecutionEndpointMetrics.list.Time(func() {
 		response, err = m.NodeExecutionManager.ListNodeExecutions(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListNodeExecutions",
-		audit.ParametersFromExecutionIdentifier(request.WorkflowExecutionId),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.nodeExecutionEndpointMetrics.list)
 	}
@@ -93,7 +69,6 @@ func (m *AdminService) ListNodeExecutions(
 func (m *AdminService) ListNodeExecutionsForTask(
 	ctx context.Context, request *admin.NodeExecutionForTaskListRequest) (*admin.NodeExecutionList, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -109,12 +84,6 @@ func (m *AdminService) ListNodeExecutionsForTask(
 	m.Metrics.nodeExecutionEndpointMetrics.listChildren.Time(func() {
 		response, err = m.NodeExecutionManager.ListNodeExecutionsForTask(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListNodeExecutionsForTask",
-		audit.ParametersFromTaskExecutionIdentifier(request.TaskExecutionId),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.nodeExecutionEndpointMetrics.listChildren)
 	}
@@ -125,7 +94,6 @@ func (m *AdminService) ListNodeExecutionsForTask(
 func (m *AdminService) GetNodeExecutionData(
 	ctx context.Context, request *admin.NodeExecutionGetDataRequest) (*admin.NodeExecutionGetDataResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -134,12 +102,6 @@ func (m *AdminService) GetNodeExecutionData(
 	m.Metrics.nodeExecutionEndpointMetrics.getData.Time(func() {
 		response, err = m.NodeExecutionManager.GetNodeExecutionData(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"GetNodeExecutionData",
-		audit.ParametersFromNodeExecutionIdentifier(request.Id),
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.nodeExecutionEndpointMetrics.getData)
 	}
