+++ pkg/rpc/adminservice/project.go
@@ -2,9 +2,6 @@ package adminservice
 
 import (
 	"context"
-	"time"
-
-	"github.com/flyteorg/flyteadmin/pkg/audit"
 
 	"github.com/flyteorg/flyteadmin/pkg/rpc/adminservice/util"
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
@@ -15,7 +12,6 @@ import (
 func (m *AdminService) RegisterProject(ctx context.Context, request *admin.ProjectRegisterRequest) (
 	*admin.ProjectRegisterResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -24,14 +20,6 @@ func (m *AdminService) RegisterProject(ctx context.Context, request *admin.Proje
 	m.Metrics.projectEndpointMetrics.register.Time(func() {
 		response, err = m.ProjectManager.CreateProject(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"RegisterProject",
-		map[string]string{
-			audit.Project: request.Project.Id,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.projectEndpointMetrics.register)
 	}
@@ -41,7 +29,6 @@ func (m *AdminService) RegisterProject(ctx context.Context, request *admin.Proje
 
 func (m *AdminService) ListProjects(ctx context.Context, request *admin.ProjectListRequest) (*admin.Projects, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -50,12 +37,6 @@ func (m *AdminService) ListProjects(ctx context.Context, request *admin.ProjectL
 	m.Metrics.projectEndpointMetrics.list.Time(func() {
 		response, err = m.ProjectManager.ListProjects(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"ListProjects",
-		map[string]string{},
-		audit.ReadOnly,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.projectEndpointMetrics.list)
 	}
@@ -67,7 +48,6 @@ func (m *AdminService) ListProjects(ctx context.Context, request *admin.ProjectL
 func (m *AdminService) UpdateProject(ctx context.Context, request *admin.Project) (
 	*admin.ProjectUpdateResponse, error) {
 	defer m.interceptPanic(ctx, request)
-	requestedAt := time.Now()
 	if request == nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Incorrect request, nil requests not allowed")
 	}
@@ -76,14 +56,6 @@ func (m *AdminService) UpdateProject(ctx context.Context, request *admin.Project
 	m.Metrics.projectEndpointMetrics.register.Time(func() {
 		response, err = m.ProjectManager.UpdateProject(ctx, *request)
 	})
-	audit.NewLogBuilder().WithAuthenticatedCtx(ctx).WithRequest(
-		"UpdateProject",
-		map[string]string{
-			audit.Project: request.Id,
-		},
-		audit.ReadWrite,
-		requestedAt,
-	).WithResponse(time.Now(), err).Log(ctx)
 	if err != nil {
 		return nil, util.TransformAndRecordError(err, &m.Metrics.projectEndpointMetrics.update)
 	}
